dil desteği eklemeleri devamı

+++ WebProject/Controllers/AdoptionController.cs
@@ -19,38 +19,47 @@ namespace WebProject.Controllers
             _context = context;
           
         }
-        
-       
         public IActionResult Index()
         {
            var list = _context.Animals.ToList();
-           return View(list);
+            if (list.Count == 0)
+            {
+                return NotFound();
+            }
+            return View(list);
         }
         
         [HttpGet]
         public IActionResult AdoptForm(int id)
         {
-            Id= id;
+            if(id == 0)
+            {
+                return NotFound();
+            }
+            Id = id;
             return View();
         }
         [HttpPost]
         public async Task<IActionResult> AdoptForm(Application a)
         {
-            Adopted adopted = new Adopted();
-            a.AnimalId= Id;
-            _context.Applications.Add(a);
-            var animal = await _context.Animals.FindAsync(Id);
-            adopted.OldId = animal.Id;
-            adopted.Breed = animal.Breed;
-            adopted.Info = "adopted";
-            adopted.Age = animal.Age;
-            _context.Adoptees.Add(adopted);
-            _context.Animals.Remove(animal);
-            _context.SaveChanges();
-            return RedirectToAction(nameof(Index));
+           
+            if (ModelState.IsValid)
+            {
+                Adopted adopted = new Adopted();
+                a.AnimalId = Id;
+                _context.Applications.Add(a);
+                var animal = await _context.Animals.FindAsync(Id);
+                adopted.OldId = animal.Id;
+                adopted.Breed = animal.Breed;
+                adopted.Info = "adopted";
+                adopted.Age = animal.Age;
+                _context.Adoptees.Add(adopted);
+                _context.Animals.Remove(animal);
+                _context.SaveChanges();
+                return RedirectToAction(nameof(Index));
+            }
+            return View();
+            
         }
-       
-
-
     }
 }
